requirements.txt
fastapi
uvicorn
prometheus-client
pytest
flake8
bandit
httpx
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-api
